requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
clip==1.0
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
decorator==5.1.1
Deprecated==1.2.14
distro==1.8.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
grpcio==1.59.3
grpcio-tools==1.59.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
imageio==2.33.0
importlib-metadata==6.9.0
importlib-resources==6.1.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
lazy_loader==0.3
llama-index==0.9.10
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
openai==1.3.7
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
portalocker==2.8.2
prompt-toolkit==3.0.41
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
qdrant-client==1.6.9
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
tifffile==2023.9.26
tiktoken==0.5.1
toml==0.10.2
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
wcwidth==0.2.12
wrapt==1.16.0
yarl==1.9.3
zipp==3.17.0
